Don't check database or provision settings at container startup
Summary:
Application configuration can be provisioned:

  - through environment
  - through a configuration file

The current image uses environment, which is good, but actually use it
to create a configuration file (which is rather bad, as it's not comprehensive).

Ideally, we should strive to allow Etherpad to directly parse environment,
but this isn't a current upstream goal.

So, to provision a comprehensive settings.json through rOPS is probably
better than tweak environment to recreate such a file each container start.

The implementation is straightforward: bypass the entrypoint,
ie don't check anymore if connection works and don't provision a settings.json.

Node so starts directly.

Test Plan: Tested on Equatower

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1775